Extract logic of optimistic todo update to a custom hook.
